batch: Group batch source modules
Batch source creation, caching, annotation, advancement, refresh, and
selector behavior currently occupy separate files at the batch package
root.

Keeping these related modules flat makes their shared responsibility harder
to find among unrelated batch operations.

This commit places the source modules under one package and retains
temporary forwarding imports while remaining callers migrate.

The next tests-only commit will mirror the source package and exercise its
nested modules directly.
