+++ src/git_stage_batch/batch/source/__init__.py
@@ -1,0 +1,1 @@
+"""Batch source creation, lookup, and refresh."""
